# packages/types/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vex Protocol API",
        "version": "4.3.2",
        "description": "REST API for the Vex encrypted chat platform. Messages are serialized using msgpack.",
        "license": {
            "name": "AGPL-3.0-or-later"
        },
        "contact": {
            "name": "Vex Protocol",
            "url": "https://github.com/vex-protocol/vex-protocol/tree/master/packages/types"
        }
    },
    "servers": [
        {
            "url": "https://api.vex.wtf",
            "description": "Production"
        },
        {
            "url": "http://localhost:16777",
            "description": "Local development"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "paths": {
...
        "/user/{id}": {
            "get": {
                "operationId": "getUser",
                "summary": "Get user profile",
                "description": "Retrieve the public profile for a user by their ID.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "lastSeen": {
                                            "type": "string",
                                            "description": "Last activity timestamp"
                                        },
                                        "userID": {
                                            "type": "string",
                                            "description": "Unique user identifier"
                                        },
                                        "username": {
                                            "type": "string",
                                            "description": "Display username"
                                        }
                                    },
                                    "required": [
                                        "lastSeen",
                                        "userID",
                                        "username"
                                    ],
                                    "description": "Public user profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/migration/challenge": {
            "post": {
                "operationId": "requestHomeserverMigrationChallenge",
                "summary": "Request a homeserver migration challenge",
                "description": "Create a short-lived challenge that the authenticated device must sign before this homeserver prepares portable account data.",
                "tags": [
                    "migration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/msgpack": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "destinationHomeserverId": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    }
                                },
                                "required": [
                                    "destinationHomeserverId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device-signing challenge issued",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "accountId": {
                                            "type": "string",
                                            "pattern": "^0x[0-9a-f]{64}$",
                                            "description": "Canonical lowercase 32-byte hex value"
                                        },
                                        "destinationHomeserverId": {
                                            "type": "string",
                                            "pattern": "^0x[0-9a-f]{64}$",
                                            "description": "Canonical lowercase 32-byte hex value"
                                        },
                                        "deviceKey": {
                                            "type": "string",
                                            "pattern": "^0x[0-9a-f]{64}$",
                                            "description": "Canonical lowercase 32-byte hex value"
                                        },
                                        "expiresAt": {
                                            "type": "integer",
                                            "exclusiveMinimum": 0,
                                            "maximum": 9007199254740991
                                        },
                                        "nonce": {
                                            "type": "string",
                                            "pattern": "^0x[0-9a-f]{64}$",
                                            "description": "Canonical lowercase 32-byte hex value"
                                        },
                                        "sourceHomeserverId": {
                                            "type": "string",
                                            "pattern": "^0x[0-9a-f]{64}$",
                                            "description": "Canonical lowercase 32-byte hex value"
                                        }
                                    },
                                    "required": [
                                        "accountId",
                                        "destinationHomeserverId",
                                        "deviceKey",
                                        "expiresAt",
                                        "nonce",
                                        "sourceHomeserverId"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Account or device is not active locally"
                    },
                    "404": {
                        "description": "Destination homeserver is not registered"
                    }
                }
            }
        },
        "/migration/prepare": {
            "post": {
                "operationId": "prepareHomeserverMigration",
                "summary": "Authorize a homeserver migration",
                "description": "Verify the device-signed challenge and retain a time-limited source transfer. Portable state is frozen only after the finalized registry route moves to the destination.",
                "tags": [
                    "migration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/msgpack": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "accountId": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    },
                                    "destinationHomeserverId": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    },
                                    "deviceKey": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    },
                                    "expiresAt": {
                                        "type": "integer",
                                        "exclusiveMinimum": 0,
                                        "maximum": 9007199254740991
                                    },
                                    "nonce": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    },
                                    "sourceHomeserverId": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    },
                                    "signature": {
                                        "type": "string",
                                        "pattern": "^[0-9a-f]{128}$"
                                    }
                                },
                                "required": [
                                    "accountId",
                                    "destinationHomeserverId",
                                    "deviceKey",
                                    "expiresAt",
                                    "nonce",
                                    "sourceHomeserverId",
                                    "signature"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Migration transfer prepared",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "expiresAt": {
                                            "type": "integer",
                                            "exclusiveMinimum": 0,
                                            "maximum": 9007199254740991
                                        },
                                        "migrationId": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "required": [
                                        "expiresAt",
                                        "migrationId"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Authorization or device signature is invalid"
                    },
                    "409": {
                        "description": "Authorization nonce conflicts with prior use"
                    }
                }
            }
        },
        "/migration/import": {
            "post": {
                "operationId": "importHomeserverMigration",
                "summary": "Import portable account state",
                "description": "Import authoritative room references, encrypted attachments, queued ciphertext, and the public avatar from the previous homeserver after the registry route is finalized.",
                "tags": [
                    "migration"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/msgpack": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "migrationId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "sourceHomeserverId": {
                                        "type": "string",
                                        "pattern": "^0x[0-9a-f]{64}$",
                                        "description": "Canonical lowercase 32-byte hex value"
                                    }
                                },
                                "required": [
                                    "migrationId",
                                    "sourceHomeserverId"
                                ],
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Portable state import result",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "avatarImported": {
                                            "type": "boolean"
                                        },
                                        "filesImported": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "mailDeferred": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "mailImported": {
                                            "type": "integer",
                                            "minimum": 0
                                        },
                                        "roomsImported": {
                                            "type": "integer",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "avatarImported",
                                        "filesImported",
                                        "mailDeferred",
                                        "mailImported",
                                        "roomsImported"
                                    ],
                                    "additionalProperties": false
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Account or device is not active locally"
                    },
                    "409": {
                        "description": "Finalized route or migration state is invalid"
                    },
                    "502": {
                        "description": "Source homeserver returned invalid data"
                    }
                }
            }
        },
...
    },
...
}